Consolidated discussion logic
